+++ src/scicloj/metamorph/ml/tools.clj
@@ -118,6 +118,27 @@
                   new-token-behaviour)))
   ))
 
+(defn keys-in
+  "Returns a sequence of all key paths in a given map using DFS walk."
+  [m]
+  (letfn [(children [node]
+            (let [v (get-in m node)]
+              (if (map? v)
+                (map (fn [x] (conj node x)) (keys v))
+                [])))
+          (branch? [node] (-> (children node) seq boolean))]
+    (->> (keys m)
+         (map vector)
+         (mapcat #(tree-seq branch? children %)))))
+
+(defn reduce-result [r result-dissoc-in-seq]
+  (reduce (fn [x y]
+            (dissoc-in x y))
+          r
+          result-dissoc-in-seq))
+
+
+
 (comment
   (def m (java.util.HashMap.))